removed unnecessary files and modularized parsing
